+++ apps/oxlint/test/fixtures/definePlugin/.oxlintrc.json
@@ -3,11 +3,6 @@
   "categories": { "correctness": "off" },
   "rules": {
     "define-plugin-plugin/create": "error",
-    "define-plugin-plugin/create-once": "error",
-    "define-plugin-plugin/create-once-before-false": "error",
-    "define-plugin-plugin/create-once-before-only": "error",
-    "define-plugin-plugin/create-once-after-only": "error",
-    "define-plugin-plugin/create-once-hooks-only": "error",
-    "define-plugin-plugin/create-once-no-hooks": "error"
+    "define-plugin-plugin/create-once": "error"
   }
 }
