--- apps/oxlint/test/fixtures/definePlugin_and_defineRule/.oxlintrc.json
@@ -1,13 +1,0 @@
-{
-  "jsPlugins": ["./plugin.ts"],
-  "categories": { "correctness": "off" },
-  "rules": {
-    "define-plugin-and-rule-plugin/create": "error",
-    "define-plugin-and-rule-plugin/create-once": "error",
-    "define-plugin-and-rule-plugin/create-once-before-false": "error",
-    "define-plugin-and-rule-plugin/create-once-before-only": "error",
-    "define-plugin-and-rule-plugin/create-once-after-only": "error",
-    "define-plugin-and-rule-plugin/create-once-hooks-only": "error",
-    "define-plugin-and-rule-plugin/create-once-no-hooks": "error"
-  }
-}
