--- apps/oxlint/test/fixtures/definePlugin_and_defineRule/options.json
@@ -1,3 +1,0 @@
-{
-  "eslint": true
-}
