+++ apps/oxlint/test/fixtures/defineRule/.oxlintrc.json
@@ -3,11 +3,6 @@
   "categories": { "correctness": "off" },
   "rules": {
     "define-rule-plugin/create": "error",
-    "define-rule-plugin/create-once": "error",
-    "define-rule-plugin/create-once-before-false": "error",
-    "define-rule-plugin/create-once-before-only": "error",
-    "define-rule-plugin/create-once-after-only": "error",
-    "define-rule-plugin/create-once-hooks-only": "error",
-    "define-rule-plugin/create-once-no-hooks": "error"
+    "define-rule-plugin/create-once": "error"
   }
 }
