--- apps/oxlint/test/fixtures/defineRule/options.json
@@ -1,3 +1,0 @@
-{
-  "eslint": true
-}
