+++ apps/oxlint/test/fixtures/eslintCompat/.oxlintrc.json
@@ -1,0 +1,13 @@
+{
+  "jsPlugins": ["./plugin.ts"],
+  "categories": { "correctness": "off" },
+  "rules": {
+    "eslint-compat-plugin/create": "error",
+    "eslint-compat-plugin/create-once": "error",
+    "eslint-compat-plugin/create-once-before-false": "error",
+    "eslint-compat-plugin/create-once-before-only": "error",
+    "eslint-compat-plugin/create-once-after-only": "error",
+    "eslint-compat-plugin/create-once-hooks-only": "error",
+    "eslint-compat-plugin/create-once-no-hooks": "error"
+  }
+}
