+++ apps/oxlint/test/fixtures/eslintCompat/files/1.js
