+++ apps/oxlint/test/fixtures/eslintCompat/files/2.js
