+++ apps/oxlint/test/fixtures/eslintCompat/options.json
